Bug: 7354
Fixed Dylan Mode fontification of block comments when they're edited. Editing
the text of a multi-line block comment removed the comment face and fontified it
as code. The problem is that the Dylan Mode region fontification code was
narrowing the buffer to the start of the edited line, preventing font-lock from
seeing and parsing the start of the comment. Now it only narrows the buffer to
the end of the interchange header, so all the Dylan code is visible.


git-svn-id: svn+ssh://svn.gwydiondylan.org/scm/svn/dylan/trunk/gwydion@11316 4f5ac42c-83e7-0310-beca-a700b25708b8